+++ website/docs/components/inline-input/examples/basic.jsx
@@ -4,8 +4,8 @@ import InlineInput from '@semcore/ui/inline-input';
 const Example = () => {
   return (
     <InlineInput onBlurBehavior="cancel" onCancel={console.log} onChange={console.log}>
-      <InlineInput.Addon>user name:</InlineInput.Addon>
-      <InlineInput.Value defaultValue="Hello world" onConfirm={console.log} />
+      <InlineInput.Addon id="label" la>user name:</InlineInput.Addon>
+      <InlineInput.Value aria-labelledby="label" defaultValue="Hello world" onConfirm={console.log} />
       <InlineInput.ConfirmControl />
       <InlineInput.CancelControl />
     </InlineInput>