+++ website/docs/data-display/area-chart/examples/area-without-data.jsx
@@ -18,7 +18,7 @@ export default () => {
     .domain([0, 10]);
 
   return (
-    <Plot data={data} scale={[xScale, yScale]} width={width} height={height} label='example of a chart describing missing data'>
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis>
         <YAxis.Ticks />
         <YAxis.Grid />