+++ website/docs/data-display/stacked-area-chart/examples/stacked-area-without-data.jsx
@@ -27,7 +27,7 @@ export default () => {
     .domain([0, 15]);
 
   return (
-    <Plot data={data} scale={[xScale, yScale]} width={width} height={height} label='example of a chart describing missing data'>
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis>
         <YAxis.Ticks />
         <YAxis.Grid />